moved Handlebar helpers to separate file
